[FIXED JENKINS-26975] Three issues in the Jenkins performance plugin are fixed, resulting an significantly better performance (Disc, CPU and memory wise) of that plugin). An elaborate description of the changes can be found on https://issues.jenkins-ci.org/browse/JENKINS-26975

+++ src/test/java/hudson/plugins/performance/IagoParserTest.java
@@ -259,7 +259,7 @@ public class IagoParserTest extends HudsonTestCase {
   
     float nbError = 0;
     for (UriReport currentReport : report.getUriReportMap().values()) {
-      nbError += Float.valueOf(currentReport.getSummarizerErrors());      
+      nbError += currentReport.getSummarizerErrors();      
     }
 
     double errorPercent = ((double) nbError) / report.getUriReportMap().size();
